Sketch out GIF picker.
// FREEBIE
